pyproject.toml
[project]
name = "agentic-facilities-maintenance"
version = "1.0.0"
description = "Agentic AI Facilities Maintenance Assistant with governance-first design and human-in-the-loop controls"
readme = "README.md"
requires-python = ">=3.11,<=3.13"

[project.urls]
Repository = "https://github.com/Dewale-A/AgenticFacilitiesMaintenance"

[tool.poetry]
name = "agentic-facilities-maintenance"
version = "1.0.0"
description = "Agentic AI Facilities Maintenance Assistant with governance-first design and human-in-the-loop controls"
authors = ["Dewale-A"]

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
crewai = ">=0.86.0"
crewai-tools = ">=0.17.0"
langchain = ">=0.3.0"
langchain-openai = ">=0.3.0"
langchain-chroma = ">=0.2.0"
chromadb = ">=0.5.0"
fastapi = ">=0.115.0"
uvicorn = ">=0.34.0"
pydantic = ">=2.0.0"
pydantic-settings = ">=2.0.0"
python-dotenv = ">=1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
